requirements.txt
fastapi
uvicorn
gradio
pydantic
openenv-core
httpx>=0.28.1